refactor: reorganize project structure by centralizing hooks and content into dedicated directories and updating imports accordingly.
