+++ src/components/pages/home/sections/VisualTransformation/index.tsx
@@ -5,9 +5,9 @@ import { useLanguage } from "@/i18n/LanguageContext";
 import SectionTitle from "@/components/common/components/SectionTitle";
 import LaptopMockup from "../../components/LaptopMockup";
 import PhoneMockup from "../../components/PhoneMockup";
-import { projects } from "../../components/ProjectData";
-import { useProjectCarousel } from "./hooks/useProjectCarousel";
-import { useDeviceDetection } from "./hooks/useDeviceDetection";
+import { projects } from "@/content/home/projectData";
+import { useProjectCarousel } from "@/hooks/ui/useProjectCarousel";
+import { useDeviceDetection } from "@/hooks/ui/useDeviceDetection";
 import DeviceModePicker from "./DeviceModePicker";
 import CarouselNavigation from "./CarouselNavigation";
 import ProjectSelector from "./ProjectSelector";
