+++ src/content/home/variants.ts
