+++ tailwind.config.ts
@@ -5,7 +5,7 @@ export default {
     "./src/pages/**/*.{js,ts,jsx,tsx}",
     "./src/components/**/*.{js,ts,jsx,tsx}",
     "./src/app/**/*.{js,ts,jsx,tsx}",
-    "./src/content/**/*.{js,ts}",
+    "./src/**/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {
